refactor : user 코드 수정
