+++ src/main/java/com/ColdPitch/domain/apicontroller/UserApiController.java
@@ -1,29 +1,20 @@
 package com.ColdPitch.domain.apicontroller;
 
 import com.ColdPitch.config.security.JwtConfig;
-import com.ColdPitch.domain.entity.dto.jwt.TokenDto;
 import com.ColdPitch.domain.entity.dto.post.PostResponseDto;
-import com.ColdPitch.domain.entity.dto.user.LoginDto;
 import com.ColdPitch.domain.entity.dto.user.UserRequestDto;
 import com.ColdPitch.domain.entity.dto.user.UserResponseDto;
-import com.ColdPitch.domain.entity.user.UserType;
 import com.ColdPitch.domain.repository.UserRepository;
 import com.ColdPitch.domain.service.UserService;
-import com.ColdPitch.jwt.JwtFilter;
-import com.ColdPitch.utils.RandomUtil;
 import com.ColdPitch.utils.SecurityUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -32,8 +23,6 @@ import java.util.List;
 @RequestMapping("/api/v1/user")
 public class UserApiController {
     private final UserService userService;
-    private final UserRepository userRepository; //더미네이터 생성으로 인해 잠시 추가 나중에 삭제할것임
-    private final JwtConfig jwtConfig;
 
     @GetMapping
     @Operation(summary = "유저 전체 조회 ( 이후에 ADMIN 권한으로 열기)")
@@ -85,49 +74,5 @@ public class UserApiController {
         List<PostResponseDto> posts = userService.getEvaluatedPostsByUserFetch(
                 SecurityUtil.getCurrentUserEmail().orElseThrow(IllegalAccessError::new));
         return ResponseEntity.status(200).body(posts);
-    }
-
-    @GetMapping("/dummy")
-    @Operation(summary = "USER 더미데이터 생성", description = "default USER, 원하는 경우 ADMIN 입력")
-    public ResponseEntity<List<UserResponseDto>> makeDummyUser(@Valid @RequestParam(defaultValue = "USER", name = "userType(Default : USER)") UserType userType, @RequestParam(defaultValue = "5", name = "size(Default : 5)") int size) {
-        List<UserResponseDto> list = new ArrayList<>();
-        for (int i = 0; i < size; i++) {
-            long rand = RandomUtil.getRandom(Integer.MAX_VALUE);
-            list.add(userService.signUpUser(new UserRequestDto("nick" + rand, "name" + rand, "password" + rand, "eamil" + rand, "phone" + rand, userType)));
-        }
-        return ResponseEntity.status(200).body(list);
-    }
-
-    @Transactional
-    @GetMapping("/testUserToken")
-    @Operation(summary = "USER 테스트 토큰 생성")
-    public ResponseEntity<String> makeDummyUserToken() {
-        UserRequestDto urd = new UserRequestDto("testNick", "testName", "testPass", "testEamil@naver.com", "010-1234-1234", UserType.USER);
-        userRepository.deleteByEmail(urd.getEmail());
-        userService.signUpUser(urd);
-        TokenDto loginResponse = userService.login(new LoginDto(urd.getEmail(), urd.getPassword()));
-
-
-        HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.add(JwtFilter.AUTHORIZATION_HEADER, "Bearer " + loginResponse.getAccessToken());
-        String cookie = "refreshToken=" + loginResponse.getRefreshToken() + "; Path=/; Max-Age=" + jwtConfig.getRefreshExpirationTime() + "; HttpOnly; readonly";
-        httpHeaders.add("Set-Cookie", cookie);
-        return new ResponseEntity<>("Bearer " + loginResponse.getAccessToken(), httpHeaders, HttpStatus.OK);
-    }
-
-    @Transactional
-    @GetMapping("/testAdminToken")
-    @Operation(summary = "ADMIN 테스트 토큰 생성")
-    public ResponseEntity<String> makeDummyAdminToken() {
-        UserRequestDto urd = new UserRequestDto("testANick", "testAName", "testAPass", "testAEamil@naver.com", "010-5678-5678", UserType.ADMIN);
-        userRepository.deleteByEmail(urd.getEmail());
-        userService.signUpUser(urd);
-        TokenDto loginResponse = userService.login(new LoginDto(urd.getEmail(), urd.getPassword()));
-
-        HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.add(JwtFilter.AUTHORIZATION_HEADER, "Bearer " + loginResponse.getAccessToken());
-        String cookie = "refreshToken=" + loginResponse.getRefreshToken() + "; Path=/; Max-Age=" + jwtConfig.getRefreshExpirationTime() + "; HttpOnly; readonly";
-        httpHeaders.add("Set-Cookie", cookie);
-        return new ResponseEntity<>("Bearer " + loginResponse.getAccessToken(), httpHeaders, HttpStatus.OK);
     }
 }
